Feature/generate schema

+++ unittests/rapidschema/meta/json_type_set_test.cpp
@@ -8,7 +8,7 @@ namespace rapidschema {
 namespace internal {
 
 using IntStringBool = JsonTypeSet<int, std::string, bool>;
-using IntFloatBool = JsonTypeSet<int, float, bool>;
+using IntFloatBool = JsonTypeSet<int, int64_t, bool>;
 
 static_assert(IntFloatBool::Unique() == false, "IntFloatBool::Unique() should return false because there are "
                                                "multiple numeric elements");
@@ -16,7 +16,7 @@ static_assert(IntFloatBool::Unique() == false, "IntFloatBool::Unique() should re
 static_assert(IntStringBool::Unique(), "IntStringBool::Unique() should return true because all elements map to "
                                        "different JsonType");
 
-static_assert(IntStringBool::Contains<float>(), "IntStringBool::Contains<float>() should return true because"
+static_assert(IntStringBool::Contains<int64_t>(), "IntStringBool::Contains<float>() should return true because"
                                                 "it contains a numeric (e.g. int) and float is numeric");
 
 static_assert(IntStringBool::Contains<std::nullptr_t>() == false, "IntStringBool::Contains<std::nullptr_t>() "
